updated interaction and level ui

+++ Waypoint/Assets/Scripts/Test/InteractionControllerCopy.cs
@@ -31,7 +31,7 @@ public class InteractionControllerCopy : MonoBehaviour {
 	private WorldSpaceObjectController popUpController;
 	public Text popUpText;
 
-	public int textVerticalOffset;
+	public float textVerticalOffset;
